🎨 improved target handling for native gates and mapped levels
