+++ src/mqt/bench/devices/__init__.py
@@ -10,11 +10,9 @@
 
 from __future__ import annotations
 
-from .handling import (
-    get_available_device_names,
-    get_available_devices,
+from .devices import get_available_device_names, get_available_devices, get_device_by_name
+from .gatesets import (
     get_available_native_gatesets,
-    get_device_by_name,
     get_native_gateset_by_name,
 )
 from .ibm import get_ibm_target
